refactor: extract auto_dispatch and twiddle_constants modules (SRP), consolidate DFT wrappers to butterflies SSOT, remove orphaned twiddle data
- Extract FftPrecision trait + macro + dispatch from kernel/mod.rs → kernel/auto_dispatch.rs
- Extract twiddle constant tables from mixed_radix/scalar/impls.rs → scalar/twiddle_constants.rs
- Consolidate winograd/radix pass-through DFT wrappers to pub(crate) use re-exports pointing to butterflies::dft SSOT
- Route impl_short_dft! macro winograd_impl arm through butterflies::dft instead of winograd
- Add ~60 composite DFT re-exports to butterflies/dft.rs, making it the SSOT for all small-DFT lookups
- Fix missing Complex64 import after twiddle extraction
- Remove 229 lines of orphaned constant data from impls.rs

+++ crates/apollo-fft-wgpu/src/infrastructure/gpu_fft/f16_plan.rs
@@ -1280,7 +1280,7 @@ mod tests {
         // Deterministic 3×3×3 field with values in [−1, 1].
         let field = ndarray::Array3::from_shape_fn((3, 3, 3), |(i, j, k)| {
             let x = (i + j * 3 + k * 7) as f32;
-            ((0.3 * x).sin() + 0.5 * (0.7 * x).cos()) as f32
+            (0.3 * x).sin() + 0.5 * (0.7 * x).cos()
         });
 
         let forward = plan.forward_native_f16(&field);
